third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/cache-storage-reporting-document.https-expected.txt

This is a testharness.js-based test.
FAIL COEP support on document. assert_equals: expected (string) "https://www1.web-platform.test:8444/images/blue.png" but got (undefined) undefined
FAIL COEP-Report-Only support on document. assert_equals: expected (string) "https://www1.web-platform.test:8444/images/blue.png" but got (undefined) undefined
Harness: the test ran to completion.
